campaign service add

+++ Backend/saleCampaign/src/main/java/com/example/saleCampaign/model/Product.java
@@ -15,10 +15,10 @@ public class Product {
 
     private String title;
 
-    private BigDecimal mrp;
+    private Double mrp;
 
     @Column(name = "current_price")
-    private BigDecimal currentPrice;
+    private Double currentPrice;
 
     private int inventory;
 }
